Update payroll.py
Completed updates required by the BitShares API. Fully tested on a production delegate node. Some questions remain concerning the delegate account, but if the payroll[ 'minimum_balance' ] is set to 0 there are no issues and all but a very small amount of the available delegate pay is distributed. If the payroll[ 'minimum_balance' ] is non-zero, the code is still fully functional without errors, however the delegate wallet balance does not appear to be increased as I expected it would be. In fact it's not clear where the funds are transfered to.